Added Password class that generates and compares password hash
