+++ SupportYourLocals.Data/CSVData.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
-using MapControl;
 using Microsoft.VisualBasic.FileIO;
 using SupportYourLocals.ExtensionMethods;
 using System.Security.Cryptography;
